라이브 트레이딩 구현 (spec/0011) + 거래소 API 검증/수정
spec/0011 전체 구현:
- ExecutionPolicy trait: SimPolicy/LivePolicy 컴파일타임 다형성
- monitor 분할: monitor_core.rs + monitor_sim.rs + monitor_live.rs
- LiveExecutor: IOC 지정가 + 3단계 긴급 청산 에스컬레이션
- BalanceTracker: RAII ReservationToken 기반 잔고 관리
- RiskManager: AtomicBool kill switch + 24h 롤링 손실 추적
- AlertService: 13개 이벤트 타입 텔레그램 알림
- LinearOrderManagement trait: Bybit spot/linear API 컴파일타임 분리
- arb-db: SQLite 기반 포지션/거래 영속화
- 바이너리 분리: main.rs(라이브) + zscore_sim.rs(시뮬레이션)

거래소 API 검증 및 수정:
- Bithumb: POST /v1/orders → /v2/orders (ord_type→order_type 필드명 변경)
- Bithumb: DELETE /v1/order+uuid → /v2/order+order_id (파라미터명 변경)
- Bithumb: v2 간소화 응답 → GET /v1/order 후속 조회 패턴 추가
- Bybit: reduceOnly 파라미터 항상 명시적 전달 (Some(reduce_only))
- Upbit/Bybit/Bithumb: 주문 관련 디버그 로깅 강화

+++ crates/arb-strategy/src/zscore/spread.rs
@@ -929,7 +929,7 @@ mod tests {
             calc.update(
                 "BTC",
                 ts + chrono::Duration::minutes(i),
-                Some(dec(138_000_000 + i as i64 * 100_000, 0)),
+                Some(dec(138_000_000 + i * 100_000, 0)),
                 1380.0,
                 Some(dec(100_000 + i * 10, 0)),
             )
@@ -1070,7 +1070,7 @@ mod tests {
             calc.update(
                 "BTC",
                 ts,
-                Some(dec(138_000_000 + i as i64 * 1000, 0)),
+                Some(dec(138_000_000 + i * 1000, 0)),
                 1380.0,
                 Some(dec(100_000 + i * 10, 0)),
             )
@@ -1088,7 +1088,7 @@ mod tests {
             calc.update(
                 "BTC",
                 ts,
-                Some(dec(138_000_000 + i as i64 * 1000, 0)),
+                Some(dec(138_000_000 + i * 1000, 0)),
                 1380.0,
                 Some(dec(100_000 + i * 10, 0)),
             )
